+++ addon/components/dropzone-widget/component.js
@@ -80,6 +80,13 @@ export default Ember.Component.extend({
             }
         });
 
+        // Dropzone.js does not have an option for disabling selecting multiple files when clicking the "upload" button.
+        // Therefore, we remove the "multiple" attribute for the hidden file input element, so that users cannot select
+        // multiple files for upload in the first place.
+        if (this.get('options.preventMultipleFiles') && this.get('clickable')){
+            Ember.$('.dz-hidden-input').removeAttr('multiple');
+        }
+
         this.set('dropzoneElement', drop);
 
         // Set osf session header