+++ index.js
@@ -101,6 +101,18 @@ module.exports = {
                 useScss: true
             };
         }
+
+        // Set options.ember-power-select.useScss to true in consuming app to use SCSS styles
+        // from ember-power-select (must import manually with @import 'ember-power-select').
+        if (!(app.options['ember-power-select'] && app.options['ember-power-select'].useScss)) {
+            // ember-power-select normally skips inclusion of the precompiled css file if
+            // the consuming app uses SASS, but using @import 'ember-power-select' in ember-osf
+            // doesn't work properly, so we just import the compiled css.
+            if (app.registry.availablePlugins['ember-cli-sass']) {
+                app.import('vendor/ember-power-select.css');
+            }
+        }
+
         return app;
     },
     treeForAddon: function(tree) {