+++ CHANGELOG.md
@@ -8,6 +8,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `lazy-options` component to enable lazy loading for `ember-power-select`
 
+### Removed
+- Ability to create new components when moving files.
+
+## [0.14.0] - 2018-02-07
+### Added
+- `osf-model.queryHasMany`, for reliable querying of hasMany relations
+- Modal to `file-browser` for moving quick files to projects
+- Actions for moving quick files to project/node in `file-browser`
+- `get-ancestor-descriptor` helper to create project/node list
+- `project-selector` component which helps user choose where to move file
+- Translations for `project-selector` component
+- Move function for `file` model
+- `ember-collapsable-panel` and `ember-power-select` packages
+- Styling on `project-selector` buttons that are selected
+
+### Changed
+- Node `addChild` model function to create public child elements
+- Import compiled css for ember-power-select even when consumin app uses SASS
+    - This can be overriden by setting ember-power-select.useSass to true in the app's options
+- Node `addChild` model method will only use defaults for undefined parameters (instead of falsey parameters)
+
+### Removed
+- ember-data-has-many-query
+    - Code that uses `model.query()` should update to `model.queryHasMany()`
+
 ## [0.13.1] - 2018-01-11
 ### Changed
 - Updated CHANGELOG to reflect 0.13.0 release
@@ -52,11 +77,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Added model tests for `review-action`
   - Customize model adapter
 - `meta.total` to `meta.total_pages` in osf-serializer
+- `hasHighlightedSubjects` path in preprint-providers model
 
 ### Fixed
 - Margins for scrollbar on `file-browser`
 - Clickability on dropzone widget
 - Handle Dropzone enable/disable properly
+- Prevent users from selecting multiple files by clicking 'Upload' button in Quickfiles
 
 ## [0.12.4] - 2017-12-04
 ### Added